:construction: Consumer is receiving on defined topics
Pushing current state of the connector. Currently the makeConsumer function takes in the configuration and setups and consumer accordingly to it.

For now this consumer is able to receive messages on the defined topics.

Added missing field to the configuration and also made the defaults more meaningful.

Temporary renamed the docker.yml to docker-compose

Signed-off by: Simon Pelczer <templum.dev@gmail.com>