:bug: :wrench: Removed the option to provide a Queue Name
As this was the cause of a bug. The Exchange is declared as direct, which implies it sends messages based on the routing key to queues that bound to it.
Having only a single queue caused the exchange to send all topics to it, which resulted in subscriber receiving messages that are not meant for them
Signed-off-by: Simon Pelczer <templum.dev@gmail.com>